+++ src/com/intellij/idea/plugin/hybris/facet/YFacetEditorTab.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -81,7 +81,11 @@ class YFacetEditorTab(
             }
         }
 
-        if (state.subModuleType == null && state.type != ModuleDescriptorType.PLATFORM && state.type != ModuleDescriptorType.CCV2) {
+        if (state.subModuleType == null
+            && state.type != ModuleDescriptorType.PLATFORM
+            && state.type != ModuleDescriptorType.CCV2
+            && state.type != ModuleDescriptorType.ANGULAR
+        ) {
             group("Settings") {
                 row {
                     checkBox("")