+++ src/com/intellij/idea/plugin/hybris/project/configurators/AngularConfigurator.kt
@@ -19,18 +19,27 @@
 package com.intellij.idea.plugin.hybris.project.configurators
 
 import com.intellij.idea.plugin.hybris.project.descriptors.ModuleDescriptor
-import com.intellij.openapi.progress.ProgressIndicator
+import com.intellij.idea.plugin.hybris.project.descriptors.impl.AngularModuleDescriptor
+import com.intellij.openapi.module.ModuleManager
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.Ref
+import com.intellij.openapi.vfs.VfsUtil
+import org.angular2.cli.Angular2ProjectConfigurator
 
-interface GroupModuleConfigurator {
+class AngularConfigurator {
 
-    fun process(
-        indicator: ProgressIndicator,
-        modulesChosenForImport: Collection<ModuleDescriptor>
-    )
+    fun configureAfterImport(project: Project, moduleDescriptors: List<ModuleDescriptor>): List<() -> Unit> = moduleDescriptors
+        .filterIsInstance<AngularModuleDescriptor>()
+        .mapNotNull {
+            val vfs = VfsUtil.findFileByIoFile(it.moduleRootDirectory, true)
+                ?: return@mapNotNull null
+            val moduleRef = ModuleManager.getInstance(project).findModuleByName(it.ideaModuleName())
+                ?.let { Ref.create(it) }
+                ?: return@mapNotNull null
 
-    fun process(
-        moduleDescriptor: ModuleDescriptor,
-        parents: MutableCollection<ModuleDescriptor>
-    )
+            {
+                Angular2ProjectConfigurator().configureProject(project, vfs, moduleRef, true)
+            }
+        }
 
 }