+++ src/com/intellij/idea/plugin/hybris/project/configurators/ConfiguratorFactory.kt
@@ -1,7 +1,7 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -53,6 +53,7 @@ class ConfiguratorFactory {
     fun getMavenConfigurator(): MavenConfigurator? = ApplicationManager.getApplication().getService(MavenConfigurator::class.java)
     fun getEclipseConfigurator(): EclipseConfigurator? = ApplicationManager.getApplication().getService(EclipseConfigurator::class.java)
     fun getGradleConfigurator(): GradleConfigurator? = ApplicationManager.getApplication().getService(GradleConfigurator::class.java)
+    fun getAngularConfigurator(): AngularConfigurator? = ApplicationManager.getApplication().getService(AngularConfigurator::class.java)
     fun getLoadedConfigurator(): LoadedConfigurator = ApplicationManager.getApplication().getService(LoadedConfigurator::class.java)
 
     fun getAntConfigurator(): AntConfigurator? =