+++ src/com/intellij/idea/plugin/hybris/project/descriptors/impl/AbstractModuleDescriptor.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -115,7 +115,7 @@ abstract class AbstractModuleDescriptor(
     override fun addSpringFile(file: String) = springFileSet.add(file)
     override fun getDirectDependencies() = directDependencies
 
-    override fun addDirectDependencies(dependencies: Set<ModuleDescriptor>) = this.directDependencies.addAll(dependencies)
+    override fun addDirectDependencies(dependencies: Collection<ModuleDescriptor>) = this.directDependencies.addAll(dependencies)
     internal open fun initDependencies(moduleDescriptors: Map<String, ModuleDescriptor>): Set<String> = emptySet()
 
     override fun toString() = "${javaClass.simpleName} {name=$name, moduleRootDirectory=$moduleRootDirectory}"