+++ src/com/intellij/idea/plugin/hybris/project/descriptors/impl/YWebSubModuleDescriptor.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -33,7 +33,7 @@ class YWebSubModuleDescriptor(
     override val subModuleDescriptorType: SubModuleDescriptorType = SubModuleDescriptorType.WEB
 ) : AbstractYSubModuleDescriptor(owner, moduleRootDirectory, name) {
 
-    override fun addDirectDependencies(dependencies: Set<ModuleDescriptor>): Boolean {
+    override fun addDirectDependencies(dependencies: Collection<ModuleDescriptor>): Boolean {
         dependencies
             .filterIsInstance<YModuleDescriptor>()
             .flatMap { it.getAllDependencies() }