+++ src/com/intellij/idea/plugin/hybris/project/services/HybrisProjectService.java
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -28,6 +29,8 @@ public interface HybrisProjectService {
     boolean isConfigModule(@NotNull File file);
 
     boolean isCCv2Module(@NotNull File file);
+
+    boolean isAngularModule(@NotNull File file);
 
     boolean isPlatformModule(@NotNull File file);