+++ src/com/intellij/idea/plugin/hybris/project/services/impl/DefaultHybrisProjectService.java
@@ -1,7 +1,7 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -46,6 +46,11 @@ public class DefaultHybrisProjectService implements HybrisProjectService {
                 || file.getAbsolutePath().contains(HybrisConstants.CCV2_JS_STOREFRONT_NAME)
             )
             && new File(file, HybrisConstants.CCV2_MANIFEST_NAME).isFile();
+    }
+
+    @Override
+    public boolean isAngularModule(@NotNull final File file) {
+        return new File(file, HybrisConstants.FILE_ANGULAR_JSON).isFile();
     }
 
     @Override